Update component servie
