+++ packages/strapi-plugin-content-type-builder/controllers/ContentTypes.js
@@ -1,17 +1,14 @@
 'use strict';
 
 const _ = require('lodash');
-const pluralize = require('pluralize');
-const fse = require('fs-extra');
-const path = require('path');
 
-const generator = require('strapi-generate');
-const { formatAttributes, convertAttributes } = require('../utils/attributes');
 const { nameToSlug } = require('../utils/helpers');
 const {
   validateContentTypeInput,
   validateUpdateContentTypeInput,
 } = require('./validation/content-type');
+
+const contentTypeService = require('../services/ContentTypes');
 
 module.exports = {
   getContentTypes(ctx) {
@@ -22,7 +19,9 @@ module.exports = {
 
         return true;
       })
-      .map(uid => formatContentType(strapi.contentTypes[uid]));
+      .map(uid =>
+        contentTypeService.formatContentType(strapi.contentTypes[uid])
+      );
 
     ctx.send({
       data: contentTypes,
@@ -38,7 +37,7 @@ module.exports = {
       return ctx.send({ error: 'contentType.notFound' }, 404);
     }
 
-    ctx.send({ data: formatContentType(contentType) });
+    ctx.send({ data: contentTypeService.formatContentType(contentType) });
   },
 
   async createContentType(ctx) {
@@ -60,11 +59,11 @@ module.exports = {
     strapi.reload.isWatching = false;
 
     try {
-      const contentType = createContentTypeSchema(body);
+      const contentType = contentTypeService.createContentTypeSchema(body);
 
-      await generateAPI(modelName, contentType);
+      await contentTypeService.generateAPI(modelName, contentType);
 
-      await generateReversedRelations({
+      await contentTypeService.generateReversedRelations({
         attributes: body.attributes,
         modelName,
       });
@@ -84,11 +83,14 @@ module.exports = {
 
     setImmediate(() => strapi.reload());
 
-    ctx.send({
-      data: {
-        uid,
+    ctx.send(
+      {
+        data: {
+          uid,
+        },
       },
-    });
+      201
+    );
   },
 
   async updateContentType(ctx) {
@@ -110,14 +112,17 @@ module.exports = {
     strapi.reload.isWatching = false;
 
     try {
-      const newSchema = updateContentTypeSchema(contentType.__schema__, body);
+      const newSchema = contentTypeService.updateContentTypeSchema(
+        contentType.__schema__,
+        body
+      );
 
-      await writeContentType({ uid, schema: newSchema });
+      await contentTypeService.writeContentType({ uid, schema: newSchema });
 
       // delete all relations directed to the updated ct except for oneWay and manyWay
-      await deleteBidirectionalRelations(contentType);
+      await contentTypeService.deleteBidirectionalRelations(contentType);
 
-      await generateReversedRelations({
+      await contentTypeService.generateReversedRelations({
         attributes: body.attributes,
         modelName: contentType.modelName,
         plugin: contentType.plugin,
@@ -128,12 +133,9 @@ module.exports = {
       } else {
         strapi.emit('didCreateContentType');
       }
-    } catch (e) {
-      strapi.log.error(e);
-      strapi.emit('didNotCreateContentType', e);
-      return ctx.badRequest(null, [
-        { messages: [{ id: 'request.error.model.write' }] },
-      ]);
+    } catch (error) {
+      strapi.emit('didNotCreateContentType', error);
+      throw error;
     }
 
     setImmediate(() => strapi.reload());
@@ -160,8 +162,8 @@ module.exports = {
 
     strapi.reload.isWatching = false;
 
-    await removeContentType(contentType);
-    await deleteAllRelations(contentType);
+    await contentTypeService.deleteAllRelations(contentType);
+    await contentTypeService.removeContentType(contentType);
 
     setImmediate(() => strapi.reload());
 
@@ -171,304 +173,4 @@ module.exports = {
       },
     });
   },
-};
-
-const deleteAllRelations = ({ modelName, plugin }) => {
-  const contentTypeUpdates = Object.keys(strapi.contentTypes).map(uid => {
-    const { __schema__ } = strapi.contentTypes[uid];
-
-    const keysToDelete = Object.keys(__schema__.attributes).filter(key => {
-      const attr = __schema__.attributes[key];
-      const target = attr.model || attr.collection;
-
-      const sameModel = target === modelName;
-      const samePluginOrNoPlugin =
-        (attr.plugin && attr.plugin === plugin) || !attr.plugin;
-
-      if (samePluginOrNoPlugin && sameModel) {
-        return true;
-      }
-
-      return false;
-    });
-
-    if (keysToDelete.length > 0) {
-      const newchema = {
-        ...__schema__,
-        attributes: _.omit(__schema__.attributes, keysToDelete),
-      };
-
-      return writeContentType({ uid, schema: newchema });
-    }
-  });
-
-  const componentUpdates = Object.keys(strapi.components).map(uid => {
-    const { __schema__ } = strapi.components[uid];
-
-    const keysToDelete = Object.keys(__schema__.attributes).filter(key => {
-      const attr = __schema__.attributes[key];
-      const target = attr.model || attr.collection;
-
-      const sameModel = target === modelName;
-      const samePluginOrNoPlugin =
-        (attr.plugin && attr.plugin === plugin) || !attr.plugin;
-
-      if (samePluginOrNoPlugin && sameModel) {
-        return true;
-      }
-
-      return false;
-    });
-
-    if (keysToDelete.length > 0) {
-      const newchema = {
-        ...__schema__,
-        attributes: _.omit(__schema__.attributes, keysToDelete),
-      };
-
-      return strapi.plugins[
-        'content-type-builder'
-      ].services.components.writeComponent({ uid, schema: newchema });
-    }
-  });
-
-  return Promise.all([...contentTypeUpdates, ...componentUpdates]);
-};
-
-const deleteBidirectionalRelations = ({ modelName, plugin }) => {
-  const updates = Object.keys(strapi.contentTypes).map(uid => {
-    const { __schema__ } = strapi.contentTypes[uid];
-
-    const keysToDelete = Object.keys(__schema__.attributes).filter(key => {
-      const attr = __schema__.attributes[key];
-      const target = attr.model || attr.collection;
-
-      const sameModel = target === modelName;
-      const samePluginOrNoPlugin =
-        (attr.plugin && attr.plugin === plugin) || !attr.plugin;
-
-      const isBiDirectionnal = _.has(attr, 'via');
-
-      if (samePluginOrNoPlugin && sameModel && isBiDirectionnal) {
-        return true;
-      }
-
-      return false;
-    });
-
-    if (keysToDelete.length > 0) {
-      const newchema = {
-        ...__schema__,
-        attributes: _.omit(__schema__.attributes, keysToDelete),
-      };
-
-      return writeContentType({ uid, schema: newchema });
-    }
-  });
-
-  return Promise.all(updates);
-};
-
-const buildReversedRelation = ({ key, attr, plugin, modelName }) => {
-  const targetAttributeOptions = {
-    via: key,
-    columnName: attr.targetColumnName,
-    plugin,
-  };
-
-  switch (attr.nature) {
-    case 'manyWay':
-    case 'oneWay':
-      return;
-    case 'oneToOne':
-    case 'oneToMany':
-      targetAttributeOptions.model = modelName;
-      break;
-    case 'manyToOne':
-      targetAttributeOptions.collection = modelName;
-      break;
-    case 'manyToMany': {
-      targetAttributeOptions.collection = modelName;
-
-      if (!targetAttributeOptions.dominant) {
-        targetAttributeOptions.dominant = true;
-      }
-      break;
-    }
-    default:
-  }
-
-  return targetAttributeOptions;
-};
-
-const generateReversedRelations = ({ attributes, modelName, plugin }) => {
-  const promises = Object.keys(attributes)
-    .filter(key => _.has(attributes[key], 'target'))
-    .map(key => {
-      const attr = attributes[key];
-      const target = strapi.contentTypes[attr.target];
-
-      const schema = _.merge({}, target.__schema__, {
-        attributes: {
-          [attr.targetAttribute]: buildReversedRelation({
-            key,
-            attr,
-            plugin,
-            modelName,
-          }),
-        },
-      });
-
-      return writeContentType({ uid: attr.target, schema });
-    });
-
-  return Promise.all(promises);
-};
-
-const removeContentType = async ({ uid }) => {
-  const { apiName, __filename__ } = strapi.contentTypes[uid];
-
-  const baseName = path.basename(__filename__, '.settings.json');
-  const apiFolder = path.join(strapi.dir, 'api', apiName);
-
-  const deleteFile = async filePath => {
-    const fileName = path.basename(filePath);
-
-    if (_.startsWith(_.toLower(fileName), _.toLower(baseName) + '.')) {
-      await fse.remove(filePath);
-    }
-
-    if (fileName === 'routes.json') {
-      const { routes } = await fse.readJSON(filePath);
-
-      const clearedRoutes = routes.filter(route => {
-        return !_.startsWith(
-          _.toLower(route.handler),
-          _.toLower(baseName) + '.'
-        );
-      });
-
-      if (clearedRoutes.length === 0) {
-        await fse.remove(filePath);
-      } else {
-        await fse.writeJSON(
-          filePath,
-          {
-            routes: clearedRoutes,
-          },
-          {
-            spaces: 2,
-          }
-        );
-      }
-    }
-  };
-
-  const recursiveRemoveFiles = async folder => {
-    const filesName = await fse.readdir(folder);
-
-    for (const fileName of filesName) {
-      const filePath = path.join(folder, fileName);
-
-      const stat = await fse.stat(filePath);
-
-      if (stat.isDirectory()) {
-        await recursiveRemoveFiles(filePath);
-      } else {
-        await deleteFile(filePath);
-      }
-    }
-
-    const files = await fse.readdir(folder);
-    if (files.length === 0) {
-      await fse.remove(folder);
-    }
-  };
-
-  await recursiveRemoveFiles(apiFolder);
-};
-
-const writeContentType = async ({ uid, schema }) => {
-  const { plugin, apiName, __filename__ } = strapi.contentTypes[uid];
-
-  let fileDir;
-  if (plugin) {
-    fileDir = `./extensions/${plugin}/models`;
-  } else {
-    fileDir = `./api/${apiName}/models`;
-  }
-
-  const filePath = path.join(strapi.dir, fileDir, __filename__);
-
-  await fse.ensureFile(filePath);
-  return fse.writeFile(filePath, JSON.stringify(schema, null, 2));
-};
-
-const formatContentType = contentType => {
-  const { uid, plugin, connection, collectionName, info } = contentType;
-
-  return {
-    uid,
-    plugin,
-    schema: {
-      name: _.get(info, 'name') || _.upperFirst(pluralize(uid)),
-      description: _.get(info, 'description', ''),
-      connection,
-      collectionName,
-      attributes: formatAttributes(contentType),
-    },
-  };
-};
-
-const createContentTypeSchema = infos => ({
-  connection:
-    infos.connection ||
-    _.get(
-      strapi,
-      ['config', 'currentEnvironment', 'database', 'defaultConnection'],
-      'default'
-    ),
-  collectionName:
-    infos.collectionName || `${_.snakeCase(pluralize(infos.name))}`,
-  info: {
-    name: infos.name,
-    description: infos.description,
-  },
-  attributes: convertAttributes(infos.attributes),
-});
-
-const updateContentTypeSchema = (old, infos) => ({
-  ...old,
-  connection: infos.connection || old.connection,
-  collectionName: infos.collectionName || old.collectionName,
-  info: {
-    name: infos.name || old.info.name,
-    description: infos.description || old.info.description,
-  },
-  // TODO: keep old params like autoMigration, private, configurable
-  attributes: convertAttributes(infos.attributes),
-});
-
-const generateAPI = (name, contentType) => {
-  // create api
-  return new Promise((resolve, reject) => {
-    const scope = {
-      generatorType: 'api',
-      id: name,
-      name,
-      rootPath: strapi.dir,
-      args: {
-        displayName: contentType.info.name,
-        description: contentType.info.description,
-        connection: contentType.connection,
-        collectionName: contentType.collectionName,
-        attributes: contentType.attributes,
-      },
-    };
-
-    generator(scope, {
-      success: () => resolve(),
-      error: err => reject(err),
-    });
-  });
 };
