+++ packages/strapi-plugin-content-type-builder/services/Components.js
@@ -5,36 +5,17 @@ const _ = require('lodash');
 const fse = require('fs-extra');
 const pluralize = require('pluralize');
 
+const contentTypeService = require('./ContentTypes');
 const { formatAttributes, convertAttributes } = require('../utils/attributes');
 const { nameToSlug } = require('../utils/helpers');
-
-/**
- * Returns a list of all available components with formatted attributes
- */
-const getComponents = () => {
-  return Object.keys(strapi.components).map(uid => {
-    return formatComponent(uid, strapi.components[uid]);
-  });
-};
-
-/**
- * Returns a component by uid
- * @param {string} uid - component's UID
- */
-const getComponent = uid => {
-  const component = strapi.components[uid];
-  if (!component) return null;
-
-  return formatComponent(uid, component);
-};
 
 /**
  * Formats a component attributes
  * @param {string} uid - string
  * @param {Object} component - strapi component model
  */
-const formatComponent = (uid, component) => {
-  const { connection, collectionName, info, category } = component;
+const formatComponent = component => {
+  const { uid, connection, collectionName, info, category } = component;
 
   return {
     uid,
@@ -58,7 +39,12 @@ const formatComponent = (uid, component) => {
 async function createComponent({ uid, infos }) {
   const schema = createSchema(infos);
 
-  await writeSchema({ uid, schema });
+  await writeSchema({
+    category: nameToSlug(infos.category),
+    name: nameToSlug(infos.name),
+    schema,
+  });
+
   return { uid };
 }
 
@@ -67,49 +53,43 @@ async function createComponent({ uid, infos }) {
  * @param {Object} component
  * @param {Object} infos
  */
-async function updateComponent({ component, newUID, infos }) {
-  const { uid, schema: oldSchema } = component;
+async function updateComponent({ component, infos }) {
+  const { uid, __schema__: oldSchema } = component;
 
   // don't update collectionName if not provided
   const updatedSchema = {
-    info: {
-      icon: infos.icon,
-      name: infos.name,
-      description: infos.description || oldSchema.description,
-    },
+    ...oldSchema,
     connection: infos.connection || oldSchema.connection,
     collectionName: infos.collectionName || oldSchema.collectionName,
+    info: {
+      name: infos.name || oldSchema.info.name,
+      icon: infos.icon || oldSchema.info.icon,
+      description: infos.description || oldSchema.info.description,
+    },
     attributes: convertAttributes(infos.attributes),
   };
 
-  if (uid !== newUID) {
-    await deleteSchema(uid);
+  await editSchema({ uid, schema: updatedSchema });
 
-    if (_.has(strapi.plugins, ['content-manager', 'services', 'components'])) {
-      await _.get(strapi.plugins, [
-        'content-manager',
-        'services',
-        'components',
-      ]).updateUID(uid, newUID);
+  if (component.category !== infos.category) {
+    const oldDir = path.join(strapi.dir, 'components', component.category);
+    const newDir = path.join(strapi.dir, 'components', infos.category);
+
+    await fse.move(
+      path.join(oldDir, component.__filename__),
+      path.join(newDir, component.__filename__)
+    );
+
+    const list = await fse.readdir(oldDir);
+    if (list.length === 0) {
+      await fse.remove(oldDir);
     }
 
-    await writeSchema({
-      uid: newUID,
-      schema: updatedSchema,
-    });
-
-    const [category] = uid.split('.');
-
-    const categoryDir = path.join(strapi.dir, 'components', category);
-    const categoryCompos = await fse.readdir(categoryDir);
-    if (categoryCompos.length === 0) {
-      await fse.remove(categoryDir);
-    }
-
-    return { uid: newUID };
+    return {
+      uid: `${infos.category}.${component.modelName}`,
+    };
   }
 
-  await writeSchema({ uid, schema: updatedSchema });
   return { uid };
 }
 
@@ -164,14 +144,27 @@ async function deleteComponent(component) {
 /**
  * Writes a component schema file
  */
-async function writeSchema({ uid, schema }) {
-  const [category, filename] = uid.split('.');
-  const categoryDir = path.join(strapi.dir, 'components', category);
+async function writeSchema({ category, name, schema }) {
+  const filePath = path.join(
+    strapi.dir,
+    'components',
+    category,
+    `${name}.json`
+  );
 
-  await fse.ensureDir(categoryDir);
+  await fse.ensureFile(filePath);
+  await fse.writeJSON(filePath, schema, { spaces: 2 });
+}
 
-  const filepath = path.join(categoryDir, `${filename}.json`);
-  await fse.writeFile(filepath, JSON.stringify(schema, null, 2));
+/**
+ * Edit a component schema file
+ */
+async function editSchema({ uid, schema }) {
+  const { category, __filename__ } = strapi.components[uid];
+  const filePath = path.join(strapi.dir, 'components', category, __filename__);
+
+  await fse.ensureFile(filePath);
+  await fse.writeJSON(filePath, schema, { spaces: 2 });
 }
 
 /**
@@ -179,156 +172,131 @@ async function writeSchema({ uid, schema }) {
  * @param {string} ui
  */
 async function deleteSchema(uid) {
-  const [category, filename] = uid.split('.');
-  await strapi.fs.removeAppFile(`components/${category}/${filename}.json`);
+  const { category, __filename__ } = strapi.components[uid];
+  await strapi.fs.removeAppFile(`components/${category}/${__filename__}`);
 }
 
 const updateComponentInModels = (oldUID, newUID) => {
-  const contentTypeService =
-    strapi.plugins['content-type-builder'].services.contenttypebuilder;
+  const contentTypeUpdates = Object.keys(strapi.contentTypes).map(uid => {
+    const { __schema__: oldSchema } = strapi.contentTypes[uid];
 
-  const updateModels = (models, { plugin } = {}) => {
-    Object.keys(models).forEach(modelKey => {
-      const model = models[modelKey];
+    const componentsToUpdate = Object.keys(oldSchema.attributes).reduce(
+      (acc, key) => {
+        if (
+          oldSchema.attributes[key].type === 'component' &&
+          oldSchema.attributes[key].component === oldUID
+        ) {
+          acc.push(key);
+        }
 
-      const attributesToModify = Object.keys(model.attributes).reduce(
-        (acc, key) => {
-          if (
-            model.attributes[key].type === 'component' &&
-            model.attributes[key].component === oldUID
-          ) {
-            acc.push(key);
-          }
+        return acc;
+      },
+      []
+    );
 
-          return acc;
-        },
-        []
-      );
-
-      const dynamicoznesToUpdate = Object.keys(model.attributes).filter(key => {
+    const dynamiczonesToUpdate = Object.keys(oldSchema.attributes).filter(
+      key => {
         return (
-          model.attributes[key].type === 'dynamiczone' &&
-          model.attributes[key].components.includes(oldUID)
+          oldSchema.attributes[key].type === 'dynamiczone' &&
+          oldSchema.attributes[key].components.includes(oldUID)
         );
-      }, []);
+      },
+      []
+    );
 
-      if (attributesToModify.length > 0) {
-        const modelJSON = contentTypeService.readModel(modelKey, {
-          plugin,
-          api: model.apiName,
-        });
+    if (componentsToUpdate.length > 0 || dynamiczonesToUpdate.length > 0) {
+      const newSchema = _.cloneDeep(oldSchema);
 
-        attributesToModify.forEach(key => {
-          modelJSON.attributes[key].component = newUID;
-        });
-
-        dynamicoznesToUpdate.forEach(key => {
-          modelJSON.attributes[key] = {
-            ...modelJSON.attributes[key],
-            components: modelJSON.attributes[key].components.map(val =>
-              val !== oldUID ? val : newUID
-            ),
-          };
-        });
-
-        contentTypeService.writeModel(modelKey, modelJSON, {
-          plugin,
-          api: model.apiName,
-        });
-      }
-    });
-  };
-
-  updateModels(strapi.models);
-
-  Object.keys(strapi.plugins).forEach(pluginKey => {
-    updateModels(strapi.plugins[pluginKey].models, { plugin: pluginKey });
-  });
-
-  Object.keys(strapi.components).forEach(uid => {
-    const component = strapi.components[uid];
-
-    const componentsToRemove = Object.keys(component.attributes).filter(key => {
-      return (
-        component.attributes[key].type === 'component' &&
-        component.attributes[key].component === oldUID
-      );
-    }, []);
-
-    if (componentsToRemove.length > 0) {
-      const newSchema = {
-        info: component.info,
-        connection: component.connection,
-        collectionName: component.collectionName,
-        attributes: component.attributes,
-      };
-
-      componentsToRemove.forEach(key => {
+      componentsToUpdate.forEach(key => {
         newSchema.attributes[key].component = newUID;
       });
 
-      writeSchema({ uid, schema: newSchema });
+      dynamiczonesToUpdate.forEach(key => {
+        newSchema.attributes[key].components = oldSchema.attributes[
+          key
+        ].components.map(val => (val !== oldUID ? val : newUID));
+      });
+
+      return contentTypeService.writeContentType({ uid, schema: newSchema });
     }
+
+    return Promise.resolve();
   });
+
+  const componentUpdates = Object.keys(strapi.components).map(uid => {
+    const { __schema__: oldSchema } = strapi.components[uid];
+
+    const componentsToUpdate = Object.keys(oldSchema.attributes).filter(key => {
+      return (
+        oldSchema.attributes[key].type === 'component' &&
+        oldSchema.attributes[key].component === oldUID
+      );
+    }, []);
+
+    if (componentsToUpdate.length > 0) {
+      const newSchema = {
+        ...oldSchema,
+      };
+
+      componentsToUpdate.forEach(key => {
+        newSchema.attributes[key].component = newUID;
+      });
+
+      return editSchema({ uid, schema: newSchema });
+    }
+
+    return Promise.resolve();
+  });
+
+  return Promise.all([...contentTypeUpdates, ...componentUpdates]);
 };
 
 const deleteComponentInModels = async componentUID => {
-  const [category] = componentUID.split('.');
-  const contentTypeService =
-    strapi.plugins['content-type-builder'].services.contenttypebuilder;
+  const component = strapi.components[componentUID];
 
-  const updateModels = (models, { plugin } = {}) => {
-    Object.keys(models).forEach(modelKey => {
-      const model = models[modelKey];
+  const contentTypeUpdates = Object.keys(strapi.contentTypes).map(uid => {
+    const { __schema__: oldSchema } = strapi.contentTypes[uid];
 
-      const componentsToRemove = Object.keys(model.attributes).filter(key => {
+    const componentsToRemove = Object.keys(oldSchema.attributes).filter(key => {
+      return (
+        oldSchema.attributes[key].type === 'component' &&
+        oldSchema.attributes[key].component === componentUID
+      );
+    }, []);
+
+    const dynamiczonesToUpdate = Object.keys(oldSchema.attributes).filter(
+      key => {
         return (
-          model.attributes[key].type === 'component' &&
-          model.attributes[key].component === componentUID
+          oldSchema.attributes[key].type === 'dynamiczone' &&
+          oldSchema.attributes[key].components.includes(componentUID)
         );
-      }, []);
+      },
+      []
+    );
 
-      const dynamicoznesToUpdate = Object.keys(model.attributes).filter(key => {
-        return (
-          model.attributes[key].type === 'dynamiczone' &&
-          model.attributes[key].components.includes(componentUID)
-        );
-      }, []);
+    if (componentsToRemove.length > 0 || dynamiczonesToUpdate.length > 0) {
+      const newSchema = _.cloneDeep(oldSchema);
 
-      if (componentsToRemove.length > 0 || dynamicoznesToUpdate.length > 0) {
-        const modelJSON = contentTypeService.readModel(modelKey, {
-          plugin,
-          api: model.apiName,
-        });
+      componentsToRemove.forEach(key => {
+        delete newSchema.attributes[key];
+      });
 
-        componentsToRemove.forEach(key => {
-          delete modelJSON.attributes[key];
-        });
+      dynamiczonesToUpdate.forEach(key => {
+        newSchema.attributes[key] = {
+          ...newSchema.attributes[key],
+          components: newSchema.attributes[key].components.filter(
+            val => val !== componentUID
+          ),
+        };
+      });
 
-        dynamicoznesToUpdate.forEach(key => {
-          modelJSON.attributes[key] = {
-            ...modelJSON.attributes[key],
-            components: modelJSON.attributes[key].components.filter(
-              val => val !== componentUID
-            ),
-          };
-        });
+      return contentTypeService.writeContentType({ uid, schema: newSchema });
+    }
 
-        contentTypeService.writeModel(modelKey, modelJSON, {
-          plugin,
-          api: model.apiName,
-        });
-      }
-    });
-  };
-
-  updateModels(strapi.models);
-
-  Object.keys(strapi.plugins).forEach(pluginKey => {
-    updateModels(strapi.plugins[pluginKey].models, { plugin: pluginKey });
+    return Promise.resolve();
   });
 
-  Object.keys(strapi.components).forEach(uid => {
+  const componentUpdates = Object.keys(strapi.components).map(uid => {
     const component = strapi.components[uid];
 
     const componentsToRemove = Object.keys(component.attributes).filter(key => {
@@ -350,25 +318,28 @@ const deleteComponentInModels = async componentUID => {
         delete newSchema.attributes[key];
       });
 
-      writeSchema({ uid, schema: newSchema });
+      return editSchema({ uid, schema: newSchema });
     }
+
+    return Promise.resolve();
   });
 
-  const categoryDir = path.join(strapi.dir, 'components', category);
-  const categoryCompos = await fse.readdir(categoryDir);
-  if (categoryCompos.length === 0) {
+  await Promise.all([...contentTypeUpdates, ...componentUpdates]);
+
+  const categoryDir = path.join(strapi.dir, 'components', component.category);
+  const list = await fse.readdir(categoryDir);
+  if (list.length === 0) {
     await fse.remove(categoryDir);
   }
 };
 
 module.exports = {
-  getComponents,
-  getComponent,
   createComponent,
   createComponentUID,
   updateComponent,
   deleteComponent,
-  writeComponent: writeSchema,
+  editSchema,
+  formatComponent,
 
   // export for testing only
   createSchema,
